pyproject.toml
[project]
name = "lkwg-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn[standard]>=0.44.0",
]
